feat(NgZone): add support for a user provided NgZone
fixes #2683
